added variable pre and post li offset, to allow double li link integration

+++ src/jokrey/utilities/ring_buffer/validation/CrashableVSRB.java
@@ -34,7 +34,7 @@ public class CrashableVSRB extends VarSizedRingBufferQueueOnly {
 
     //Returns whether the element was added
     public boolean append(byte[] e, CrashPoint crash) {
-        int lieSize = lieSize(e);
+        long lieSize = lieSize(e);
         long newDrEnd;
 
         long nextWriteStart = drStart;//start writing at dirty region
